refactor: change JsonProperty to JsonNaming
