+++ src/main/java/gift/kakaoApi/dto/token/KakaoTokenResponse.java
@@ -1,0 +1,21 @@
+package gift.kakaoApi.dto.token;
+
+import com.fasterxml.jackson.databind.PropertyNamingStrategies;
+import com.fasterxml.jackson.databind.annotation.JsonNaming;
+
+@JsonNaming(PropertyNamingStrategies.SnakeCaseStrategy.class)
+public record KakaoTokenResponse(
+
+
+    String tokenType,
+
+    String accessToken,
+
+    String expiresIn,
+
+    String refreshToken,
+
+    String refreshTokenExpiresIn
+){
+    
+}
